luffy_code/luffyapi/requirements.txt
amqp==2.6.1
asgiref==3.4.1
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
cached-property==1.5.2
celery==4.4.0
certifi==2024.2.2
charset-normalizer==2.0.12
click==7.1.2
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.25
django-ckeditor==6.3.0
django-cors-headers==3.10.1
django-crispy-forms==1.13.0
django-filter==2.4.0
django-formtools==2.4
django-import-export==2.7.1
django-js-asset==2.0.0
django-redis==5.4.0
django-reversion==4.0.2
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
dnspython==2.2.1
et-xmlfile==1.1.0
eventlet==0.33.3
future==1.0.0
greenlet==2.0.2
httplib2==0.22.0
idna==3.7
importlib-metadata==4.8.3
importlib-resources==5.4.0
kombu==4.6.11
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
packaging==21.3
Pillow==8.4.0
prompt-toolkit==3.0.36
PyJWT==1.7.1
PyMySQL==1.0.2
pyparsing==3.1.2
pytz==2024.1
PyYAML==6.0.1
redis==4.3.6
requests==2.27.1
ronglian-sms-sdk==1.0
six==1.16.0
sqlparse==0.4.4
tablib==3.1.0
typing-extensions==4.1.1
urllib3==1.26.18
vine==1.3.0
wcwidth==0.2.13
xlrd==2.0.1
XlsxWriter==3.2.0
xlwt==1.3.0
zipp==3.6.0
